requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
reportlab>=4.0.0
python-pptx>=0.6.21
openpyxl>=3.1.0
anthropic>=0.20.0
python-dotenv>=1.0.0
